Finish basic CMS

+++ admin/includes/view_all_comments.php
@@ -45,7 +45,7 @@
             
             echo "<td><a href='comments.php?approved={$comment_id}'>Approve</a> </td>";
             echo "<td><a href='comments.php?unapproved={$comment_id}'>Unapprove</a> </td>";
-            echo "<td><a href='comments.php?delete={$comment_id}'>Delete</a> </td>";
+            echo "<td><a onClick=\" javascript: return confirm('Are you sure you want to delete?');\" href='comments.php?delete={$comment_id}'>Delete</a> </td>";
             echo "</tr>";
 
         }
