+++ admin/includes/view_all_posts.php
@@ -92,7 +92,7 @@ echo "<td><img width='100' src='../images/$post_image' alt='image'></td>";
             echo "<td>{$post_status}</td>";
             echo "<td><a href='../post.php?p_id={$post_id}'>View</a></td>";
             echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a> </td>";
-            echo "<td><a href='posts.php?delete={$post_id}'>Delete</a> </td>";
+            echo "<td><a onClick=\" javascript: return confirm('Are you sure you want to delete?');\" href='posts.php?delete={$post_id}'>Delete</a> </td>";
             echo "</tr>";
 
         }
